♻ refactor time settings
