+++ src/lib/tools/time.ts
@@ -1,9 +1,50 @@
 import dayjs from 'dayjs';
 import utc from 'dayjs/plugin/utc';
-import timeZone from 'dayjs/plugin/timezone';
-
 dayjs.extend(utc);
+
+import timeZone from 'dayjs/plugin/timezone';
 dayjs.extend(timeZone);
+
+import relativeTime from 'dayjs/plugin/relativeTime';
+const dayJSConfig = {
+    thresholds: [
+        { l: 's', r: 1 },
+        { l: 'ss', r: 59, d: 'second' },
+        { l: 'm', r: 1 },
+        { l: 'mm', r: 59, d: 'minute' },
+        { l: 'h', r: 1 },
+        { l: 'hh', r: 23, d: 'hour' },
+        { l: 'd', r: 1 },
+        { l: 'dd', r: 29, d: 'day' },
+        { l: 'M', r: 1 },
+        { l: 'MM', r: 11, d: 'month' },
+        { l: 'y' },
+        { l: 'yy', d: 'year' }
+    ]
+};
+dayjs.extend(relativeTime, dayJSConfig);
+
+import updateLocale from 'dayjs/plugin/updateLocale';
+dayjs.extend(updateLocale);
+
+dayjs.updateLocale('en', {
+    relativeTime: {
+        future: 'in %s',
+        past: '%s ago',
+        s: '%d second',
+        ss: '%d seconds',
+        m: '%d minute',
+        mm: '%d minutes',
+        h: '%d hour',
+        hh: '%d hours',
+        d: '%d day',
+        dd: '%d days',
+        M: '%d month',
+        MM: '%d months',
+        y: '%d year',
+        yy: '%d years'
+    }
+});
 
 export function timeNow(
     timezone: string,
@@ -59,8 +100,8 @@ export function timeNow(
 
 export function convertTime(time: number): string {
     const now = dayjs();
-    const timeTaken = dayjs.unix(now.unix() - Math.round(time / 1000));
-    return `${timeTaken.fromNow(true)}`;
+    // const timeTaken = dayjs.unix(now.unix() - Math.round(time / 1000));
+    return `${dayjs.unix(now.unix() - Math.round(time / 1000)).fromNow(true)}`;
 }
 
 export function uptime(): string {
